scribe/standalone/src/main/java/com/karuslabs/scribe/standalone/resolvers/SingleResolver.java
/*
 * The MIT License
 *
 * Copyright 2020 Karus Labs.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 */
package com.karuslabs.scribe.standalone.resolvers;

import com.karuslabs.scribe.standalone.Resolver;

import java.util.Set;
import javax.annotation.processing.Messager;
import javax.lang.model.element.Element;

import static javax.tools.Diagnostic.Kind.ERROR;


public abstract class SingleResolver extends Resolver {
    
    protected String annotation;
    
    
    public SingleResolver(Messager messager, String annotation) {
        super(messager);
        this.annotation = annotation;
    }
    
    
    @Override
    protected boolean check(Set<? extends Element> elements) {
        var single = elements.size() <= 1;
        if (!single) {
            for (var element : elements) {
                messager.printMessage(
                    ERROR, 
                    "Invalid number of @" + annotation + " annotations, plugin must contain only one @" + annotation + " annotation", 
                    element
                );
            }
        }
        
        return single;
    }

}
